+++ test/test_layers_geom.py
@@ -3,7 +3,7 @@ import numpy as np
 import os
 from .utils import ASSETS_PATH
 import tensorflow as tf
-from kgcnn.graph.adj import get_angle_indices
+from kgcnn.graph.methods import get_angle_indices
 from kgcnn.layers.geom import NodeDistanceEuclidean, EdgeAngle, NodePosition
 from kgcnn.literature.DimeNetPP._dimenet_conv import SphericalBasisLayer
 from kgcnn.layers.geom import BesselBasisLayer